Add files via upload
This module contains three functions for working with discrete-time signals using the NumPy library.

The first function is called invert_signal and applies time and/or amplitude inversion to a discrete-time signal. The function takes a NumPy array as input, as well as two optional Boolean arguments for inverting the signal in time and/or amplitude. The function returns the inverted signal as a NumPy array.

The second function is called shift_signal and allows shifting a discrete-time signal in samples and/or amplitude. The function takes a NumPy array as input, as well as an integer for the number of samples to shift and a float for the amplitude value to shift. Two optional Boolean arguments allow specifying whether to shift the signal in time and/or amplitude. The function returns the shifted signal as a NumPy array.

The third function is called operate_signal and applies the selected mathematical operation between two signals of equal length. The function takes two NumPy arrays as input and a string indicating the mathematical operation to perform. Options are "add" for addition, "mult" for multiplication, and "convo" for convolution. The function returns the result of applying the selected operator to the input signals as a NumPy array.